+++ packages/manager/apps/container/src/container/legacy/server-sidebar/SidebarItemRenderer.tsx
@@ -1,9 +1,25 @@
 import React, { useCallback, useState } from 'react';
-import { Spinner } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
 import { useLegacyContainer } from '@/container/legacy/context';
 import { SidebarMenuItem } from './sidebarMenu';
 import style from './index.module.scss';
+import {
+  OsdsSpinner,
+  OsdsLink,
+  OsdsText,
+  OsdsIcon,
+} from '@ovhcloud/ods-stencil/components/react';
+import {
+  OdsSpinnerSize,
+  OdsIconName,
+  OdsHTMLAnchorElementTarget,
+  OdsHTMLAnchorElementRel,
+  OdsTextLevel,
+  OdsIconSize,
+  OdsTextSize,
+} from '@ovhcloud/ods-core';
+
+import { OdsThemeColorIntent } from '@ovhcloud/ods-theming';
 
 function ServerSidebarSearchField({
   item,
@@ -46,104 +62,128 @@ export default function ServerSidebarItemRenderer({
   const { setIsResponsiveSidebarMenuOpen } = useLegacyContainer();
   const hasToggle = item.loader || item.subItems?.length > 0;
   const updateHandler = useCallback(() => onSearch(item), [item, onSearch]);
+  const [isClicked, setIsClicked] = useState(false);
   let itemRender = null;
+
+  let linkColor = OdsThemeColorIntent.text;
+  if (item.isSelected) {
+    linkColor = OdsThemeColorIntent.primary;
+  }
+
+  const handleClick = () => {
+    setIsClicked(!isClicked);
+    onToggle(item);
+  };
   if ('searchField' in item) {
     itemRender = (
       <ServerSidebarSearchField item={item} onUpdate={updateHandler} />
     );
   } else if (item.isNoResultField) {
-    itemRender = (
-      <div>
-        {t('sidebar_no_result')}
-      </div>
-    );
+    itemRender = <div>{t('sidebar_no_result')}</div>;
   } else if (hasToggle) {
-    let icon = 'oui-icon-chevron-right';
+
+    const chevron = item.isOpen ? OdsIconName.CHEVRON_DOWN : OdsIconName.CHEVRON_RIGHT;
+    let iconOrSpinner;
+
     if (item.loadingError) {
-      icon = 'oui-icon-error-circle';
-    } else if (item.isOpen) {
-      icon = 'oui-icon-chevron-down';
+      iconOrSpinner = (
+        <OsdsIcon
+          name={OdsIconName.ERROR_CIRCLE}
+          size={OdsIconSize.sm}
+          color={OdsThemeColorIntent.text}
+        />
+
+      );
+    } else if (item.isLoading) {
+      iconOrSpinner = <OsdsSpinner size={OdsSpinnerSize.sm} />;
+    } else {
+      iconOrSpinner = (
+        <OsdsIcon
+          name={chevron}
+          size={OdsIconSize.xs}
+          color={OdsThemeColorIntent.text}
+          className={style.spinnerLoadingItem}
+        />
+      );
     }
+
     itemRender = (
-      <button
-        type="button"
-        disabled={item.isLoading}
-        onClick={() => onToggle(item)}
-        {...{
-          'aria-busy': !!item.isLoading,
-        }}
-        title={item.isOpen ? '' : t('sidebar_load_services')}
-        className={`${item.depth > 0 ? style.subButton : ''} ${
-          item.isSelected ? style.selectedButton : ''
-        }`}
-      >
-        {item.isLoading && (
-          <span className={style.menuItemIcon} aria-hidden="true">
-            <Spinner
-              as="span"
-              animation="border"
-              size="sm"
-              role="status"
-              aria-hidden="true"
-            />
-          </span>
-        )}
-        {!item.isLoading && (
-          <span className={style.menuItemIcon} aria-hidden="true">
-            <i className={`oui-icon ${icon}`}></i>
-          </span>
-        )}
-        {item.icon && (
-          <span className={style.menuItemIcon} aria-hidden="true">
+      <button className={style.transparentButton} onClick={handleClick}>
+        <span className={style.itemLineAlign}>
+          {iconOrSpinner}
+          <span className={style.iconPadding} aria-hidden="true">
             {item.icon}
+            <span className={style.textPadding}>{item.depth === 0 ? (
+              <OsdsText
+                level={OdsTextLevel.heading}
+                color={OdsThemeColorIntent.text}
+                size={OdsTextSize._200}
+              >
+                {item.label}
+              </OsdsText>
+            ) : (
+              <OsdsText
+                level={OdsTextLevel.button}
+                color={OdsThemeColorIntent.text}
+                size={OdsTextSize._300}
+              >
+                {item.label}
+              </OsdsText>
+            )}</span>
           </span>
-        )}
-        <span>{item.label}</span>
+        </span>
       </button>
     );
   } else {
-    const externalTarget = item.isExternal
+    const additionalLinkProps = item.isExternal
       ? {
-        target: '_blank',
-        rel: 'noopener',
+        target: OdsHTMLAnchorElementTarget._blank,
+        rel: OdsHTMLAnchorElementRel.noopener,
       }
       : {};
     itemRender = (
-      <a
-        style={{
-          paddingLeft: item.depth === 0 ? '1.75rem' : '0',
-          fontWeight: item.isSelected || item.depth === 0 ? 700 : 400,
-          color: item.isSelected ? '#2859c0' : '#113f6d',
-        }}
-        href={item.href || '#'}
-        onClick={() => setIsResponsiveSidebarMenuOpen(false)}
-        {...externalTarget}
-        title={item.title || t('sidebar_access_dashboard')}
-      >
-        {item.icon && (
-          <span className={style.menuItemIcon} aria-hidden="true">
-            {item.icon}
-          </span>
-        )}
-        <span>{item.label}</span>
-        {item?.badge && (
-          <span
-            className={`oui-badge oui-badge_s oui-badge_${item.badge} ${style.menuBadge}`}
+
+      <span className={style.itemLink}>
+        <span className={style.menuItemIcon} aria-hidden="true">
+          {item.icon}
+        </span>
+        <OsdsLink
+          {...additionalLinkProps}
+          role="treeitem"
+          color={linkColor}
+          href={item.href || '#'}
+          onClick={() => {
+            setIsResponsiveSidebarMenuOpen(false);
+            handleClick();
+          }}
+          className={style.link}
+        >
+          {item.depth === 0 ? <OsdsText
+            level={OdsTextLevel.heading}
+            color={OdsThemeColorIntent.text}
+            size={OdsTextSize._200}
           >
-            {item.badge}
-          </span>
-        )}
-      </a>
+            <span>{item.label}</span>
+          </OsdsText> : <span className={item.isSelected ? style.linkTextSelected : style.linkText}>{item.label}</span>}
+          {item?.badge && (
+            <span
+              className={`oui-badge oui-badge_s oui-badge_${item.badge} ${style.menuBadge}`}
+            >
+              {item.badge}
+            </span>
+          )}
+        </OsdsLink>
+
+      </span>
     );
   }
   return (
-    <div className={`${style.menuItem} ${item.depth > 0 ? 'pl-2' : ''}`}>
+    <div className={`${style.menuItem} ${item.depth > 0 ? 'pl-2' : ''} ${item.isOpen ? style.sticky : style.noSticky}`}>
       <div className={item.depth > 0 ? style.menuItemBorder : ''}>
         <div
           className={item.isSelected ? style.menuItemSelected : ''}
           style={{
-            paddingLeft: `${item.depth * 1.75 -
-              (item.depth === 1 ? 0.5 : 0)}rem`,
+            paddingLeft: `${item.depth * 1.5 - (item.depth === 1 ? 0.5 : 0)}rem`,
           }}
         >
           {itemRender}