+++ packages/manager/apps/dedicated/client/app/license/order/options/plesk/component.js
@@ -1,0 +1,13 @@
+import controller from './controller';
+import template from './template.html';
+
+export default {
+  bindings: {
+    options: '<',
+    version: '<',
+    agoraEnabled: '<',
+    onOptionsChange: '&',
+  },
+  controller,
+  template,
+};