+++ packages/manager/core/application/src/ovh-tracking.tsx
@@ -119,9 +119,12 @@ export default function OvhTracking() {
   const ovhTrackSelectOption = (event) => {
     const element = event.target as HTMLElement;
     const closestWithTracking = element.closest(`[${'data-tracking'}]`);
-    const trackingValue = closestWithTracking.getAttribute('data-tracking');
-    if (trackingValue) {
-      ovhTrackingSendClick(trackingValue);
+
+    if (closestWithTracking) {
+      const trackingValue = closestWithTracking.getAttribute('data-tracking');
+      if (trackingValue) {
+        ovhTrackingSendClick(trackingValue);
+      }
     }
   };