+++ packages/manager/core/ovh-product-icons/icons/webIcons.ts
@@ -1,0 +1,15 @@
+import {
+  Hosting,
+  Office365,
+  Sharepoint,
+  Exchange,
+} from '../assets-icon-svg/icons';
+
+const webIcons = {
+  Hosting,
+  Office365,
+  Sharepoint,
+  Exchange,
+};
+
+export default webIcons;