+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/ArrowUpIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const ArrowUpIcon = createIcon({
-  displayName: 'ArrowUpIcon',
-  viewBox: '0 0 24 24',
-  d: 'M12 22.414a1 1 0 0 1-1-1l.001-15.586-4.294 4.293a1 1 0 0 1-1.414-1.414L12 2l.7.7.015.015 5.992 5.992a1 1 0 0 1-1.414 1.414L13 5.831v15.582a1 1 0 0 1-1 1z',
-});