+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/BinIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const BinIcon = createIcon({
-  displayName: 'BinIcon',
-  viewBox: '0 0 24 24',
-  d: 'M15 5h4a1 1 0 0 1 0 2H5a1 1 0 0 1 0-2h4c0-1.857.386-1.99.883-2h4.234c.497.01.883.143.883 2zm-5 4a1 1 0 0 1 1 1v7a1 1 0 0 1-2 0v-7a1 1 0 0 1 1-1zm8-1a1 1 0 0 1 1 1v11c0 1.219-.781 2-2 2H7c-1.219 0-2-.781-2-2V9a1 1 0 0 1 2 0v11h10V9a1 1 0 0 1 1-1zm-4 1a1 1 0 0 1 1 1v7a1 1 0 0 1-2 0v-7a1 1 0 0 1 1-1z',
-});