+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/ExternalLinkIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const ExternalLinkIcon = createIcon({
-  displayName: 'ExternalLinkIcon',
-  viewBox: '0 0 24 24',
-  d: 'M12.657 2a1 1 0 1 0 0 2l6.072.002-11.022 11.02a1 1 0 0 0 0 1.414c.391.39 1.024.39 1.415 0l11.021-11.02v6.07a.999.999 0 1 0 1.999 0V2h-9.485zM3 2.051a1 1 0 0 0-1 1v18c0 .551.448 1 1 1h18a1 1 0 0 0 1-1v-3.975a1 1 0 1 0-2 0v2.975H4v-16h3.014a1 1 0 1 0 0-2H3z',
-});