+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/HamburgerIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const HamburgerIcon = createIcon({
-  displayName: 'HamburgerIcon',
-  viewBox: '0 0 24 24',
-  d: 'M18 16a1 1 0 0 1 0 2H6a1 1 0 0 1 0-2h12zm0-5a1 1 0 0 1 0 2H6a1 1 0 0 1 0-2h12zm0-5a1 1 0 0 1 0 2H6a1 1 0 1 1 0-2h12z',
-});