+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/PlusIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const PlusIcon = createIcon({
-  displayName: 'PlusIcon',
-  viewBox: '0 0 24 24',
-  d: 'M12 3a1 1 0 0 1 1 1v7h7a1 1 0 0 1 0 2h-7v7a1 1 0 0 1-2 0v-7H4a1 1 0 0 1 0-2h7V4a1 1 0 0 1 1-1z',
-});