+++ packages/manager/modules/emailpro/src/domain/dkim-autoconfig/emailpro-domain-dkim-autoconfig.controller.js
@@ -1,40 +1,60 @@
-import { DKIM_CONFIGURATION_GUIDE } from './emailpro-domain-dkim-autoconfig.constants';
+import {
+  DKIM_CONFIGURATION_GUIDE,
+  DKIM_CONFIGURATION_GUIDE_NO_OVH,
+} from './emailpro-domain-dkim-autoconfig.constants';
+import { DKIM_STATUS, DKIM_STATUS_TEXT } from '../emailpro-domain.constants';
 
-export default class EmailProDomainDkimAutoconfigCtrl {
+import DkimAutoConfigurator from './dkim-auto-configurator';
+
+export default class EmailProDomainDkimAutoconfigCtrl extends DkimAutoConfigurator {
   /* @ngInject */
   constructor(
     $q,
     $scope,
-    $stateParams,
-    $translate,
-    coreConfig,
-    EmailPro,
     EmailProDomains,
+    navigation,
+    $translate,
+    $stateParams,
+    coreConfig,
+    $state, // see DkimAutoConfigurator
   ) {
+    super();
     this.services = {
       $q,
       $scope,
-      $stateParams,
-      $translate,
-      EmailPro,
       EmailProDomains,
+      navigation,
+      $translate,
+      $stateParams,
+      $state,
     };
 
+    this.serviceType = 'emailpro';
     this.domain = $scope.currentActionData.domain;
-    this.dkimStatus = $scope.currentActionData.dkimStatus;
-    this.GLOBAL_DKIM_STATUS =
-      $scope.currentActionData.constant.GLOBAL_DKIM_STATUS;
-    this.services.$scope.configDkim = () => this.configDkim();
-
+    this.DKIM_STATUS = DKIM_STATUS;
     this.dkimGuideLink =
       DKIM_CONFIGURATION_GUIDE[coreConfig.getUser().ovhSubsidiary] ||
       DKIM_CONFIGURATION_GUIDE.DEFAULT;
+    this.dkimGuideLinkNoOvh =
+      DKIM_CONFIGURATION_GUIDE_NO_OVH[coreConfig.getUser().ovhSubsidiary] ||
+      DKIM_CONFIGURATION_GUIDE_NO_OVH.DEFAULT;
 
     this.init();
   }
 
   init() {
     this.loading = true;
+    const {
+      dkimDiag: { state, errorCode, message },
+    } = this.domain;
+    this.dkimStatus = state;
+    if (state === DKIM_STATUS.ERROR) {
+      this.dkimErrorMessage = message;
+    }
+
+    // Vars for DKIM configuration inside modal stepper
+    this.initializeDkimConfiguratorNoOvh();
+
     this.services.EmailProDomains.getDnsSettings(
       this.services.$stateParams.productId,
       this.domain.name,
@@ -43,6 +63,14 @@ export default class EmailProDomainDkimAutoconfigCtrl {
         (data) => {
           this.domainDiag = data;
           this.hideConfirmButton = this.hideConfirmButton();
+          this.dkimForNoOvhCloud =
+            this.dkimStatus === DKIM_STATUS.TO_CONFIGURE &&
+            !this.domainDiag.isOvhDomain;
+          this.bodyText = this.getBodyText(
+            state,
+            this.dkimGuideLink,
+            errorCode,
+          );
           this.loading = false;
         },
         (failure) => {
@@ -62,48 +90,127 @@ export default class EmailProDomainDkimAutoconfigCtrl {
 
   hideConfirmButton() {
     return (
-      this.dkimStatus === this.GLOBAL_DKIM_STATUS.NOT_CONFIGURED &&
-      this.domainDiag.isOvhDomain
+      this.domainDiag.isOvhDomain &&
+      [
+        DKIM_STATUS.TO_CONFIGURE,
+        DKIM_STATUS.ACTIVE,
+        DKIM_STATUS.DISABLED,
+      ].includes(this.dkimStatus)
     );
   }
 
-  configDkim() {
-    this.services.EmailProDomains.getDkimSelector(
+  onFinishDkim() {
+    switch (this.dkimStatus) {
+      case DKIM_STATUS.TO_CONFIGURE:
+        return this.configDkim();
+      case DKIM_STATUS.ACTIVE:
+        return this.deactivateDkim();
+      case DKIM_STATUS.DISABLED:
+        return this.activateDkim();
+      default:
+        console.error('Invalid DKIM status:', this.dkimStatus);
+        return null;
+    }
+  }
+
+  activateDkim() {
+    const dkimSelectors = this.domain.dkim;
+
+    const promise = this.services.EmailProDomains.enableDkim(
       this.services.$stateParams.productId,
       this.domain.name,
-    )
-      .then((dkimSelectors) => {
-        const promises = dkimSelectors.map((dkimSelector, index) => {
-          return this.services.EmailProDomains.postDkim(
-            this.services.$stateParams.productId,
-            this.domain.name,
-            {
-              selectorName: dkimSelector,
-              autoEnableDKIM: index === 0,
-              configureDkim: true,
-            },
-          );
-        });
-        return this.services.$q.all(promises);
-      })
+      dkimSelectors[0].selectorName,
+    );
+
+    this.handleDkimOperationResponse(promise);
+  }
+
+  deactivateDkim() {
+    const dkimSelectors = this.domain.dkim;
+
+    const promise = this.services.EmailProDomains.disableDkim(
+      this.services.$stateParams.productId,
+      this.domain.name,
+      dkimSelectors[0].status === 'inProduction'
+        ? dkimSelectors[0].selectorName
+        : dkimSelectors[1].selectorName,
+    );
+    this.handleDkimOperationResponse(promise);
+  }
+
+  getDkimSelectorForCurrentState() {
+    return this.services.EmailProDomains.getDkimSelector(
+      this.services.$stateParams.productId,
+      this.domain.name,
+    ).catch(({ data }) => {
+      this.writeError('emailpro_tab_domain_diagnostic_dkim_error', data);
+      this.leaveDkimConfigurator();
+      return this.services.$q.reject(data);
+    });
+  }
+
+  postDkim(selectors) {
+    return selectors.map((dkimSelector, index) => {
+      return this.services.EmailProDomains.postDkim(
+        this.services.$stateParams.productId,
+        this.domain.name,
+        {
+          selectorName: dkimSelector,
+          autoEnableDKIM: index === 0,
+          configureDkim: !this.dkimForNoOvhCloud,
+        },
+      );
+    });
+  }
+
+  configDkim() {
+    const promise = this.getDkimSelectorForCurrentState();
+    promise.then((dkimSelectors) => {
+      const promises = this.postDkim(dkimSelectors);
+      return this.services.$q.all(promises);
+    });
+    this.handleDkimOperationResponse(promise);
+  }
+
+  handleDkimOperationResponse(promise) {
+    let reload = false;
+    promise
       .then(() => {
-        this.services.$scope.setMessage(
-          this.services.$translate.instant(
-            'emailpro_tab_domain_diagnostic_dkim_activation_success',
-          ),
-          { status: 'success' },
+        reload = true;
+        this.writeSuccess(
+          'emailpro_tab_domain_diagnostic_dkim_activation_success',
         );
       })
-      .catch(() => {
-        this.services.$scope.setMessage(
-          this.services.$translate.instant(
-            'emailpro_tab_domain_diagnostic_dkim_activation_error',
-          ),
-          { status: 'error' },
-        );
+      .catch(({ data }) => {
+        this.writeError('emailpro_tab_domain_diagnostic_dkim_error', data);
       })
       .finally(() => {
-        this.services.$scope.resetAction();
+        this.leaveDkimConfigurator(reload);
       });
+  }
+
+  getBodyText(status, url, errorCode) {
+    let translationKey;
+    if (status === DKIM_STATUS.TO_CONFIGURE) {
+      translationKey = this.domainDiag.isOvhDomain
+        ? 'emailpro_tab_domain_diagnostic_dkim_activation_ovhcloud'
+        : 'emailpro_tab_domain_diagnostic_dkim_activation_no_ovhcloud';
+    } else {
+      translationKey = DKIM_STATUS_TEXT[status];
+    }
+    return translationKey
+      ? this.services.$translate.instant(translationKey, {
+          url,
+          errorCode,
+        })
+      : '';
+  }
+
+  getTitleForWizard() {
+    const titleKey =
+      DKIM_STATUS.DISABLED === this.dkimStatus ? 'activation' : 'error';
+    return this.services.$translate.instant(
+      `emailpro_tab_domain_diagnostic_dkim_title_${titleKey}`,
+    );
   }
 }