+++ packages/manager/modules/pci/src/projects/project/instances/instance/instance.controller.js
@@ -6,16 +6,20 @@ export default class PciInstanceController {
     CucCloudMessage,
     ovhManagerRegionService,
     PciProjectsProjectInstanceService,
+    PciProject,
   ) {
     this.$translate = $translate;
     this.coreConfig = coreConfig;
     this.CucCloudMessage = CucCloudMessage;
     this.ovhManagerRegionService = ovhManagerRegionService;
     this.PciProjectsProjectInstanceService = PciProjectsProjectInstanceService;
+    this.PciProject = PciProject;
   }
 
   $onInit() {
     this.loadMessages();
+    this.globalRegionsUrl = this.PciProject.getDocumentUrl('GLOBAL_REGIONS');
+    this.localZoneUrl = this.PciProject.getDocumentUrl('LOCAL_ZONE');
   }
 
   displayBillingActionButton() {