+++ packages/manager/apps/catalog/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.2.0...@ovh-ux/manager-catalog-app@1.2.1) (2024-01-22)
+
+**Note:** Version bump only for package @ovh-ux/manager-catalog-app
+
+
+
+
+
+# [1.2.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.1.5...@ovh-ux/manager-catalog-app@1.2.0) (2024-01-08)
+
+
+### Features
+
+* **catalog:** app chip to display filter ([#10230](https://github.com/ovh/manager/issues/10230)) ([8bfb12e](https://github.com/ovh/manager/commit/8bfb12ea3e9f3f7fdc4e0a96d090582a4ef537a4))
+
+
+
+
+
+## [1.1.5](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.1.4...@ovh-ux/manager-catalog-app@1.1.5) (2024-01-04)
+
+**Note:** Version bump only for package @ovh-ux/manager-catalog-app
+
+
+
+
+
 ## [1.1.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-catalog-app@1.1.3...@ovh-ux/manager-catalog-app@1.1.4) (2023-12-19)
 
 **Note:** Version bump only for package @ovh-ux/manager-catalog-app