+++ packages/manager/apps/iam/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.17](https://github.com/ovh/manager/compare/@ovh-ux/manager-iam-app@2.0.16...@ovh-ux/manager-iam-app@2.0.17) (2024-01-22)
+
+**Note:** Version bump only for package @ovh-ux/manager-iam-app
+
+
+
+
+
 ## [2.0.16](https://github.com/ovh/manager/compare/@ovh-ux/manager-iam-app@2.0.15...@ovh-ux/manager-iam-app@2.0.16) (2023-12-19)
 
 **Note:** Version bump only for package @ovh-ux/manager-iam-app